Split TileSet/TileFactory/ParseHelpers
