+++ bin/demo-map3.tmx
@@ -18,7 +18,7 @@
 17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,
 17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,
 17,17,17,34,35,36,17,17,17,17,17,17,17,17,17,17,17,17,17,17,
-17,17,17,43,14,45,17,17,17,17,17,17,17,17,17,17,17,17,17,17,
+17,17,17,43,17,45,17,17,17,17,17,17,17,17,17,17,17,17,17,17,
 17,17,17,52,53,54,17,17,17,17,17,17,17,17,17,17,17,17,17,17,
 17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,
 17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,
