+++ src/tiled-map-loader.cpp
@@ -24,6 +24,7 @@
 #include "RenderModel.hpp"
 #include "Texture.hpp"
 #include "TileSet.hpp"
+#include "TileFactory.hpp"
 
 #include <common/StringUtil.hpp>
 
@@ -36,38 +37,11 @@ namespace {
 
 using namespace cul::exceptions_abbr;
 
-using tinyxml2::XMLDocument;
 using TeardownTask = MapLoader::TeardownTask;
 using MapSide = MapEdgeLinks::Side;
 using TileRange = MapEdgeLinks::TileRange;
 
 } // end of <anonymous> namespace
-
-TiXmlIter::TiXmlIter(): el(nullptr), name(nullptr) {}
-
-TiXmlIter::TiXmlIter(const TiXmlElement * el_, const char * name_):
-    el(el_), name(name_) {}
-
-TiXmlIter & TiXmlIter::operator ++ ()
-    { el = el->NextSiblingElement(name); return *this; }
-
-bool TiXmlIter::operator != (const TiXmlIter & rhs) const
-    { return el != rhs.el; }
-
-const TiXmlElement & TiXmlIter::operator * () const
-    { return *el; }
-
-XmlRange::XmlRange(const TiXmlElement * el_, const char * name_):
-    m_beg(el_ ? el_->FirstChildElement(name_) : nullptr, name_) {}
-
-XmlRange::XmlRange(const TiXmlElement & el_, const char * name_):
-    m_beg(el_.FirstChildElement(name_), name_) {}
-
-TiXmlIter XmlRange::begin() const { return m_beg;       }
-
-TiXmlIter XmlRange::end()   const { return TiXmlIter{}; }
-
-// ----------------------------------------------------------------------------
 
 Tuple<SharedPtr<LoaderTask>, SharedPtr<TeardownTask>> MapLoader::operator ()
     (const Vector2I & map_offset)
@@ -144,7 +118,7 @@ Tuple<SharedPtr<LoaderTask>, SharedPtr<TeardownTask>> MapLoader::operator ()
     static constexpr const std::size_t k_no_idx = std::string::npos;
     for (auto & [idx, future] : m_pending_tilesets) {
         if (!future->is_ready()) continue;
-        XMLDocument document;
+        TiXmlDocument document;
         auto contents = future->retrieve();
         document.Parse(contents.c_str());
         m_tilesets[idx]->load_information(*m_platform, *document.RootElement());
@@ -253,7 +227,7 @@ TileRange to_range
 bool is_colon(char c) { return c == ':'; }
 
 /* private */ void MapLoader::do_content_load(std::string && contents) {
-    XMLDocument document;
+    TiXmlDocument document;
     if (document.Parse(contents.c_str()) != tinyxml2::XML_SUCCESS) {
         // ...idk
         throw RtError{"Problem parsing XML x.x"};
